refactor: 프로젝트 관리 게스트 컴포넌트 정리
